Rename and move ScrollLink and themeUtils
