+++ src/components/work/carousel/carousel-controls.js
@@ -5,7 +5,7 @@ import { css } from 'emotion';
 import { FaAngleLeft, FaAngleRight, FaPlayCircle, FaPauseCircle } from 'react-icons/lib/fa';
 
 import type { Project } from '.';
-import themeUtils from '../../themeUtils';
+import themeUtils from '../../theme-utils';
 import { keyboardHandler } from '../../../utils/helpers';
 
 const ControlsFlex = styled('div')`
