+++ src/components/work/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Element } from 'react-scroll';
 import styled from 'react-emotion';
 
-import themeUtils from '../themeUtils';
+import themeUtils from '../theme-utils';
 import Carousel from './carousel';
 
 const WorkHeader = styled('h1')`
