+++ src/templates/BlogPost.js
@@ -6,7 +6,7 @@ import styled from 'react-emotion';
 import { FaAngleLeft } from 'react-icons/lib/fa';
 import 'prismjs/themes/prism.css';
 import Layout from '../layouts/main';
-import themeUtils from '../components/themeUtils';
+import themeUtils from '../components/theme-utils';
 import QuoteCard from '../components/base-components/quote-card';
 import './BlogPost.css';
